Extract useConnectActivators hook to a separate file
