+++ src/fullPageBookingForm/index.ts
@@ -1,4 +1,8 @@
 import "ts-polyfill";
+import "@fontsource/montserrat/400.css";
+import "@fontsource/montserrat/500.css";
+import "@fontsource/montserrat/600.css";
+import "@fontsource/montserrat/700.css";
 import { App, AppProps } from "./Components/App";
 import { mountComponent } from "../Utils/mount";
